+++ src/main/java/game/StoryInitializer.java
@@ -1,21 +1,17 @@
 package game;
 
-import engine.story.StoryManager;
-import engine.story.Scene;
+import engine.story.model.Story;
+import engine.story.controller.StoryManager;
 import game.stories.Test;
-import game.stories.TestStory;
 
 public class StoryInitializer {
+
     public static void setup(StoryManager storyManager) {
 
-        storyManager.addScene(Test.createScenes());
-        storyManager.setActiveScene(Test.getStartScene());
+        // Story -> Day -> Section -> Scene -> Block
+        Story story = Test.createStory();
+        storyManager.setStory(story);
 
-//        storyManager.addScene(TestStory.createScenes());
-//        storyManager.setActiveScene(TestStory.getStartScene());
-    }
-
-    public boolean isEndScene(Scene scene) {
-        return scene.isEndingScene();
+        // storyManager.goToScene(Test.getStartSceneId());
     }
 }